Working on getting Vagrant VMs spun up with Opscode Community MongoDB
I can (somewhat) successfully get the Vagrant VM m1 spun up as a single-node MongoDB server.
I am unable to get the Vagrant VM r1 spun up as a member of a replica-set because it would
seem the mongodb/definitions/mongodb.rb file requires search to determine replicaset nodes.
Using chef_solo this will not work.  I am attempting to code around this by inserting the
replicaset members as a Hash but this is not working.